refactor: extract cells
