+++ apps/antalmanac/src/components/RightPane/SectionTable/SectionTableBody/SectionTableBodyRow.tsx
@@ -1,475 +1,25 @@
-import {
-    Box,
-    Button,
-    Chip,
-    Popover,
-    TableCell,
-    TableRow,
-    Theme,
-    Tooltip,
-    Typography,
-    useMediaQuery,
-} from '@material-ui/core';
-import { withStyles } from '@material-ui/core/styles';
-import { ClassNameMap, Styles } from '@material-ui/core/styles/withStyles';
-import { AASection, WebsocSectionEnrollment, WebsocSectionMeeting, CourseDetails } from '@packages/antalmanac-types';
-import classNames from 'classnames';
-import { Fragment, memo, useCallback, useEffect, useMemo, useState } from 'react';
-import { Link } from 'react-router-dom';
+import { TableRow, useTheme } from '@mui/material';
+import { AASection, CourseDetails } from '@packages/antalmanac-types';
+import { memo, useCallback, useEffect, useMemo, useState } from 'react';
 
-import { MOBILE_BREAKPOINT } from '../../../../globals';
-import GradesPopup from '../GradesPopup';
-import { OpenSpotAlertPopoverProps } from '../OpenSpotAlertPopover';
 import { SectionActionCell } from '../cells/action';
-import restrictionsMapping from '../static/restrictionsMapping.json';
 
-import analyticsEnum, { logAnalytics } from '$lib/analytics';
-import { Grades } from '$lib/grades';
-import { clickToCopy } from '$lib/helpers';
-import locationIds from '$lib/location_ids';
+
+import { CourseCodeCell } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/CourseCodeCell';
+import { DayAndTimeCell } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/DayAndTimeCell';
+import { EnrollmentCell } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/EnrollmentCell';
+import { GpaCell } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/GpaCell';
+import { InstructorsCell } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/InstructorsCell';
+import { LocationsCell } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/LocationsCell';
+import { RestrictionsCell } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/RestrictionsCell';
+import { DetailsCell } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/SectionDetailsCell';
+import { StatusCell } from '$components/RightPane/SectionTable/SectionTableBody/SectionTableBodyCells/StatusCell';
 import AppStore from '$stores/AppStore';
 import { useColumnStore, type SectionTableColumn } from '$stores/ColumnStore';
 import { useHoveredStore } from '$stores/HoveredStore';
-import { usePreviewStore, useTimeFormatStore, useThemeStore } from '$stores/SettingsStore';
-import { useTabStore } from '$stores/TabStore';
-import { formatTimes } from '$stores/calendarizeHelpers';
+import { usePreviewStore, useThemeStore } from '$stores/SettingsStore';
 
-const styles: Styles<Theme, object> = (theme) => ({
-    sectionCode: {
-        display: 'inline-flex',
-        cursor: 'pointer',
-        '&:hover': {
-            cursor: 'pointer',
-        },
-        alignSelf: 'center',
-    },
-    row: {
-        '&:nth-of-type(odd)': {
-            backgroundColor: theme.palette.action.hover,
-        },
-    },
-    cell: {},
-    link: {
-        textDecoration: 'underline',
-        cursor: 'pointer',
-    },
-    mapLink: {
-        cursor: 'pointer',
-        background: 'none !important',
-        border: 'none',
-        padding: '0 !important',
-        fontSize: '0.85rem', // Not sure why this is not inherited
-        textDecoration: 'none',
-    },
-    paper: {
-        padding: theme.spacing(),
-    },
-    button: { padding: '6px' },
-    open: {
-        color: '#00c853',
-    },
-    waitl: {
-        color: '#1c44b2',
-    },
-    full: {
-        color: '#e53935',
-    },
-    Act: { color: '#c87137' },
-    Col: { color: '#ff40b5' },
-    Dis: { color: '#ff6e00' },
-    Fld: { color: '#1ac805' },
-    Lab: { color: '#1abbe9' },
-    Lec: { color: '#d40000' },
-    Qiz: { color: '#8e5c41' },
-    Res: { color: '#ff2466' },
-    Sem: { color: '#2155ff' },
-    Stu: { color: '#179523' },
-    Tap: { color: '#8d2df0' },
-    Tut: { color: '#ffc705' },
-    popoverText: {
-        cursor: 'pointer',
-    },
-    codeCell: {
-        width: '8%',
-    },
-    // statusCell: {
-    //     width: '9%',
-    // },
-});
-
-const NoPaddingTableCell = withStyles({
-    sizeSmall: { padding: '0px 0px 0px 0px' },
-})(TableCell);
-
-interface CourseCodeCellProps {
-    classes: ClassNameMap;
-    sectionCode: string;
-}
-
-const CourseCodeCell = withStyles(styles)((props: CourseCodeCellProps) => {
-    const isDark = useThemeStore((store) => store.isDark);
-
-    const { classes, sectionCode } = props;
-
-    const [isHovered, setIsHovered] = useState(false);
-
-    const handleMouseEnter = () => {
-        setIsHovered(true);
-    };
-
-    const handleMouseLeave = () => {
-        setIsHovered(false);
-    };
-
-    return (
-        <NoPaddingTableCell className={`${classes.cell} ${classes.codeCell}`}>
-            <Tooltip title="Click to copy course code" placement="bottom" enterDelay={150}>
-                <Chip
-                    onClick={(event) => {
-                        clickToCopy(event, sectionCode);
-                        logAnalytics({
-                            category: analyticsEnum.classSearch.title,
-                            action: analyticsEnum.classSearch.actions.COPY_COURSE_CODE,
-                        });
-                    }}
-                    className={classes.sectionCode}
-                    label={sectionCode}
-                    onMouseEnter={handleMouseEnter}
-                    onMouseLeave={handleMouseLeave}
-                    style={{
-                        color: isHovered ? (isDark ? 'gold' : 'blueviolet') : '',
-                    }}
-                    size="small"
-                />
-            </Tooltip>
-        </NoPaddingTableCell>
-    );
-});
-
-type SectionType = 'Act' | 'Col' | 'Dis' | 'Fld' | 'Lab' | 'Lec' | 'Qiz' | 'Res' | 'Sem' | 'Stu' | 'Tap' | 'Tut';
-
-interface SectionDetailCellProps {
-    classes: ClassNameMap;
-    sectionType: SectionType;
-    sectionNum: string;
-    units: number;
-}
-
-const SectionDetailsCell = withStyles(styles)((props: SectionDetailCellProps) => {
-    const { classes, sectionType, sectionNum, units } = props;
-    const isMobileScreen = useMediaQuery(`(max-width: ${MOBILE_BREAKPOINT})`);
-
-    return (
-        <NoPaddingTableCell className={classes.cell} style={isMobileScreen ? { textAlign: 'center' } : {}}>
-            <Box className={classes[sectionType]}>{sectionType}</Box>
-            <Box>
-                {!isMobileScreen && <>Sec: </>}
-                {sectionNum}
-            </Box>
-            <Box>
-                {!isMobileScreen && <>Units: </>}
-                {units}
-            </Box>
-        </NoPaddingTableCell>
-    );
-});
-
-interface InstructorsCellProps {
-    classes: ClassNameMap;
-    instructors: string[];
-}
-
-const InstructorsCell = withStyles(styles)((props: InstructorsCellProps) => {
-    const { classes, instructors } = props;
-
-    const getLinks = (professorNames: string[]) => {
-        return professorNames.map((profName, index) => {
-            if (profName !== 'STAFF') {
-                const lastName = profName.substring(0, profName.indexOf(','));
-                return (
-                    <Box key={profName}>
-                        <a
-                            href={`https://www.ratemyprofessors.com/search/professors/1074?q=${lastName}`}
-                            target="_blank"
-                            rel="noopener noreferrer"
-                        >
-                            {profName}
-                        </a>
-                    </Box>
-                );
-            } else {
-                return <Box key={profName + index}> {profName} </Box>; // The key should be fine as we're not changing ['STAFF, 'STAFF']
-            }
-        });
-    };
-
-    return <NoPaddingTableCell className={classes.cell}>{getLinks(instructors)}</NoPaddingTableCell>;
-});
-
-async function getGpaData(deptCode: string, courseNumber: string, instructors: string[]) {
-    const namedInstructors = instructors.filter((instructor) => instructor !== 'STAFF');
-
-    // Get the GPA of the first instructor of this section where data exists
-    for (const instructor of namedInstructors) {
-        const grades = await Grades.queryGrades(deptCode, courseNumber, instructor, false);
-        if (grades?.averageGPA) {
-            return {
-                gpa: grades.averageGPA.toFixed(2).toString(),
-                instructor: instructor,
-            };
-        }
-    }
-
-    return undefined;
-}
-
-interface GPACellProps {
-    deptCode: string;
-    courseNumber: string;
-    instructors: string[];
-}
-
-function GPACell(props: GPACellProps) {
-    const isDark = useThemeStore((store) => store.isDark);
-
-    const { deptCode, courseNumber, instructors } = props;
-
-    const [gpa, setGpa] = useState('');
-
-    const [instructor, setInstructor] = useState('');
-
-    const [anchorEl, setAnchorEl] = useState<Element>();
-
-    const handleClick = useCallback((event: React.MouseEvent<HTMLElement>) => {
-        setAnchorEl((currentAnchorEl) => (currentAnchorEl ? undefined : event.currentTarget));
-    }, []);
-
-    const hideDistribution = useCallback(() => {
-        setAnchorEl(undefined);
-    }, []);
-
-    useEffect(() => {
-        getGpaData(deptCode, courseNumber, instructors)
-            .then((data) => {
-                if (data) {
-                    setGpa(data.gpa);
-                    setInstructor(data.instructor);
-                }
-            })
-            .catch(console.log);
-    }, [deptCode, courseNumber, instructors]);
-
-    return (
-        <NoPaddingTableCell>
-            <Button
-                style={{
-                    color: isDark ? 'dodgerblue' : 'blue',
-                    padding: 0,
-                    minWidth: 0,
-                    fontWeight: 400,
-                    fontSize: '1rem',
-                }}
-                onClick={handleClick}
-                variant="text"
-            >
-                {gpa}
-            </Button>
-            <Popover
-                open={Boolean(anchorEl)}
-                onClose={hideDistribution}
-                anchorEl={anchorEl}
-                anchorOrigin={{ vertical: 'bottom', horizontal: 'left' }}
-                transformOrigin={{ vertical: 'top', horizontal: 'left' }}
-                disableRestoreFocus
-            >
-                <GradesPopup
-                    deptCode={deptCode}
-                    courseNumber={courseNumber}
-                    instructor={instructor}
-                    isMobileScreen={useMediaQuery(`(max-width: ${MOBILE_BREAKPOINT}`)}
-                />
-            </Popover>
-        </NoPaddingTableCell>
-    );
-}
-
-interface LocationsCellProps {
-    classes: ClassNameMap;
-    meetings: WebsocSectionMeeting[];
-    courseName: string; // Used in map pin popup
-}
-
-const LocationsCell = withStyles(styles)((props: LocationsCellProps) => {
-    const isDark = useThemeStore((store) => store.isDark);
-
-    const { classes, meetings } = props;
-
-    const { setActiveTab } = useTabStore();
-
-    const focusMap = useCallback(() => {
-        setActiveTab(2);
-    }, [setActiveTab]);
-
-    return (
-        <NoPaddingTableCell className={classes.cell}>
-            {meetings.map((meeting) => {
-                return !meeting.timeIsTBA ? (
-                    meeting.bldg.map((bldg) => {
-                        const [buildingName = ''] = bldg.split(' ');
-                        const buildingId = locationIds[buildingName];
-                        return (
-                            <Fragment key={meeting.timeIsTBA + bldg}>
-                                <Link
-                                    className={classes.mapLink}
-                                    to={`/map?location=${buildingId}`}
-                                    onClick={focusMap}
-                                    color={isDark ? 'dodgerblue' : 'blue'}
-                                >
-                                    {bldg}
-                                </Link>
-                                <br />
-                            </Fragment>
-                        );
-                    })
-                ) : (
-                    <Box>{'TBA'}</Box>
-                );
-            })}
-        </NoPaddingTableCell>
-    );
-});
-
-interface SectionEnrollmentCellProps {
-    classes: ClassNameMap;
-    numCurrentlyEnrolled: WebsocSectionEnrollment;
-    maxCapacity: number;
-
-    /**
-     * This is a string because sometimes it's "n/a"
-     */
-    numOnWaitlist: string;
-
-    /**
-     * This is a string because numOnWaitlist is a string. I haven't seen this be "n/a" but it seems possible and I don't want it to break if that happens.
-     */
-    numNewOnlyReserved: string;
-}
-
-const SectionEnrollmentCell = withStyles(styles)((props: SectionEnrollmentCellProps) => {
-    const { classes, numCurrentlyEnrolled, maxCapacity, numOnWaitlist, numNewOnlyReserved } = props;
-
-    return (
-        <NoPaddingTableCell className={classes.cell}>
-            <Box>
-                <Box>
-                    <strong>
-                        {numCurrentlyEnrolled.totalEnrolled} / {maxCapacity}
-                    </strong>
-                </Box>
-                {numOnWaitlist !== '' && <Box>WL: {numOnWaitlist}</Box>}
-                {numNewOnlyReserved !== '' && <Box>NOR: {numNewOnlyReserved}</Box>}
-            </Box>
-        </NoPaddingTableCell>
-    );
-});
-
-interface RestrictionsCellProps {
-    classes: ClassNameMap;
-    restrictions: string;
-}
-
-const RestrictionsCell = withStyles(styles)((props: RestrictionsCellProps) => {
-    const { classes, restrictions } = props;
-
-    const parseRestrictions = (restrictionCode: string) => {
-        return restrictionCode.split(' ').map((code, index) => {
-            if (code !== 'and' && code !== 'or') {
-                return (
-                    <Fragment key={index}>
-                        {restrictionsMapping[code as keyof typeof restrictionsMapping]}
-                        <br />
-                    </Fragment>
-                );
-            }
-            return null;
-        });
-    };
-
-    return (
-        <NoPaddingTableCell className={classes.cell}>
-            <Box>
-                <Tooltip title={<Typography>{parseRestrictions(restrictions)}</Typography>}>
-                    <Typography>
-                        <a
-                            href="https://www.reg.uci.edu/enrollment/restrict_codes.html"
-                            target="_blank"
-                            rel="noopener noreferrer"
-                        >
-                            {restrictions}
-                        </a>
-                    </Typography>
-                </Tooltip>
-            </Box>
-        </NoPaddingTableCell>
-    );
-});
-
-interface DayAndTimeCellProps {
-    classes: ClassNameMap;
-    meetings: WebsocSectionMeeting[];
-}
-
-const DayAndTimeCell = withStyles(styles)((props: DayAndTimeCellProps) => {
-    const { classes, meetings } = props;
-
-    const { isMilitaryTime } = useTimeFormatStore();
-
-    return (
-        <NoPaddingTableCell className={classes.cell}>
-            {meetings.map((meeting) => {
-                if (meeting.timeIsTBA) {
-                    return <Box key={meeting.timeIsTBA.toString()}>TBA</Box>;
-                }
-
-                if (meeting.startTime && meeting.endTime) {
-                    const timeString = formatTimes(meeting.startTime, meeting.endTime, isMilitaryTime);
-
-                    return <Box key={meeting.timeIsTBA + meeting.bldg[0]}>{`${meeting.days} ${timeString}`}</Box>;
-                }
-            })}
-        </NoPaddingTableCell>
-    );
-});
-
-interface StatusCellProps extends OpenSpotAlertPopoverProps {
-    term: string;
-}
-
-const StatusCell = withStyles(styles)((props: StatusCellProps) => {
-    // const { term, sectionCode, courseTitle, courseNumber, status, classes } = props;
-    const { status, classes } = props;
-
-    // TODO: Implement course notification when Anteater API has the functionality, according to #473
-    // if (term === getDefaultTerm().shortName && (status === 'NewOnly' || status === 'FULL')) {
-    //     return (
-    //         <NoPaddingTableCell className={`${classes[status.toLowerCase()]} ${classes.cell}`}>
-    //             <OpenSpotAlertPopover
-    //                 courseTitle={courseTitle}
-    //                 courseNumber={courseNumber}
-    //                 status={status}
-    //                 sectionCode={sectionCode}
-    //             />
-    //         </NoPaddingTableCell>
-    //     )
-    return (
-        <NoPaddingTableCell className={`${classes[status.toLowerCase()]} ${classes.cell} ${classes.statusCell}`}>
-            {status}
-        </NoPaddingTableCell>
-    );
-});
-
-interface SectionTableBodyProps {
-    classes: ClassNameMap;
+interface SectionTableBodyRowProps {
     section: AASection;
     courseDetails: CourseDetails;
     term: string;
@@ -483,19 +33,20 @@ interface SectionTableBodyProps {
 const tableBodyCells: Record<SectionTableColumn, React.ComponentType<any>> = {
     action: SectionActionCell,
     sectionCode: CourseCodeCell,
-    sectionDetails: SectionDetailsCell,
+    sectionDetails: DetailsCell,
     instructors: InstructorsCell,
-    gpa: GPACell,
+    gpa: GpaCell,
     dayAndTime: DayAndTimeCell,
     location: LocationsCell,
-    sectionEnrollment: SectionEnrollmentCell,
+    sectionEnrollment: EnrollmentCell,
     restrictions: RestrictionsCell,
     status: StatusCell,
 };
 
-const SectionTableBodyRow = memo((props: SectionTableBodyProps) => {
-    const { classes, section, courseDetails, term, allowHighlight, scheduleNames, scheduleConflict } = props;
+export const SectionTableBodyRow = memo((props: SectionTableBodyRowProps) => {
+    const { section, courseDetails, term, allowHighlight, scheduleNames, scheduleConflict } = props;
 
+    const theme = useTheme();
     const isDark = useThemeStore((store) => store.isDark);
     const activeColumns = useColumnStore((store) => store.activeColumns);
     const previewMode = usePreviewStore((store) => store.previewMode);
@@ -535,33 +86,53 @@ const SectionTableBodyRow = memo((props: SectionTableBodyProps) => {
     }, [updateHighlight]);
 
     const computedRowStyle = useMemo(() => {
-        /* allowHighlight is always false on CourseRenderPane and always true on AddedCoursePane */
-        const computedAddedCourseStyle = allowHighlight
-            ? isDark
-                ? { background: '#b0b04f' }
-                : { background: '#fcfc97' }
-            : {};
-        const computedScheduleConflictStyle = scheduleConflict
-            ? isDark
-                ? { background: '#121212', opacity: '0.6' }
-                : { background: '#a0a0a0', opacity: '1' }
-            : {};
-
         if (addedCourse) {
+            /* allowHighlight is always false on CourseRenderPane and always true on AddedCoursePane */
+            const computedAddedCourseStyle = allowHighlight
+                ? isDark
+                    ? { backgroundColor: '#b0b04f' }
+                    : { backgroundColor: '#fcfc97' }
+                : {};
+
             return computedAddedCourseStyle;
         }
 
         if (scheduleConflict) {
+            const computedScheduleConflictStyle = scheduleConflict
+                ? isDark
+                    ? { backgroundColor: '#121212', opacity: '0.6' }
+                    : { backgroundColor: '#a0a0a0', opacity: '1' }
+                : {};
+
             return computedScheduleConflictStyle;
         }
 
         return {};
     }, [allowHighlight, isDark, scheduleConflict, addedCourse]);
 
+    console.log({
+        '&:nth-of-type(odd)': {
+            backgroundColor: theme.palette.action.hover,
+        },
+        computedRowStyle,
+    });
+
+    /**
+     *  '&:nth-of-type(odd)': {
+                    backgroundColor: theme.palette.action.hover,
+                },
+                computedRowStyle],
+     */
     return (
         <TableRow
-            classes={{ root: classes.row }}
-            className={classNames(classes.tr)}
+            /**
+             * CSS errors occur when combining the `nth-of-type` selector with the computed styling, so it's split into two separate props
+             */
+            sx={{
+                '&:nth-of-type(odd)': {
+                    backgroundColor: theme.palette.action.hover,
+                },
+            }}
             style={computedRowStyle}
             onMouseEnter={handleMouseEnter}
             onMouseLeave={handleMouseLeave}
@@ -579,7 +150,7 @@ const SectionTableBodyRow = memo((props: SectionTableBodyProps) => {
                             scheduleConflict={scheduleConflict}
                             scheduleNames={scheduleNames}
                             {...section}
-                            sectionType={section.sectionType as SectionType}
+                            sectionType={section.sectionType}
                             maxCapacity={parseInt(section.maxCapacity, 10)}
                             units={parseFloat(section.units)}
                             courseName={`${courseDetails.deptCode} ${courseDetails.courseNumber}`}
@@ -592,5 +163,3 @@ const SectionTableBodyRow = memo((props: SectionTableBodyProps) => {
 });
 
 SectionTableBodyRow.displayName = 'SectionTableBodyRow';
-
-export default withStyles(styles)(SectionTableBodyRow);
